# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Code4Me V2 API",
    "description": "The complete API for Code4Me V2",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/completion/feedback": {
      "post": {
        "tags": [
          "Completion"
        ],
        "summary": "Submit Completion Feedback",
        "description": "Submit feedback on a generated completion.\n\nThis endpoint validates the user's session and project tokens against Redis,\nverifies the user's authorization to provide feedback on the given query,\nand enqueues asynchronous tasks to update generation status and optionally\nsave ground truth feedback.\n\nParameters:\n- feedback: The feedback data submitted by the user, including acceptance\n  status and optional ground truth.\n- app: Dependency-injected application instance providing DB and Redis access.\n- session_token: Session cookie used to authenticate the user session.\n- project_token: Project cookie used to authorize project-level access.\n\nReturns:\n- JsonResponseWithStatus: A response indicating success or describing errors,\n  including authorization failures, missing records, or internal errors.",
        "operationId": "submit_completion_feedback_api_completion_feedback_post",
        "parameters": [
          {
            "name": "session_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Session Token"
            }
          },
          {
            "name": "project_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Project Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackCompletion"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompletionFeedbackPostResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/InvalidOrExpiredSessionToken"
                    },
                    {
                      "$ref": "#/components/schemas/InvalidOrExpiredProjectToken"
                    }
                  ],
                  "title": "Response 401 Submit Completion Feedback Api Completion Feedback Post"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoAccessToProvideFeedbackError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerationNotFoundError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackRecordingError"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/completion/{query_id}": {
      "get": {
        "tags": [
          "Completion"
        ],
        "summary": "Get Completions By Query",
        "description": "Retrieve code completions associated with a specific query ID.\n\nThis endpoint validates the user's session token and authorization,\nensures the requested query exists and belongs to the requesting user,\nand then fetches all the associated completion generations from the database.\n\nParameters:\n- query_id (UUID): The unique identifier for the meta query whose completions are requested.\n- app (App, dependency): The application instance providing database and Redis access.\n- session_token (str, cookie): The session token cookie for user authentication.\n\nReturns:\n- JsonResponseWithStatus: JSON response with HTTP status and either the completions data\n  or an error response detailing the failure reason.\n\nPossible responses:\n- 200: Successfully retrieved completions for the given query ID.\n- 401: Session token is invalid, expired, or missing.\n- 403: User does not have access rights to the requested query.\n- 404: The requested query ID does not exist.\n- 422, 429: Various client errors (validation or rate limiting).\n- 500: Internal server error while retrieving completions.",
        "operationId": "get_completions_by_query_api_completion__query_id__get",
        "parameters": [
          {
            "name": "query_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Query Id"
            }
          },
          {
            "name": "session_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Session Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompletionPostResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidOrExpiredSessionToken"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoAccessToGetQueryError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryNotFoundError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetrieveCompletionsError"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/chat/get/{page_number}": {
      "get": {
        "tags": [
          "Chat"
        ],
        "summary": "Get Chat History",
        "description": "Retrieve a page of chat history for a specific project associated with the current session.\n\nParameters:\n- app (App): Dependency-injected FastAPI application instance.\n- page_number (int): The page number of the chat history to retrieve.\n- session_token (str): Session token stored in a cookie.\n- project_token (str): Project token stored in a cookie.\n\nReturns:\n- JsonResponseWithStatus: Paginated chat history or an error response.",
        "operationId": "get_chat_history_api_chat_get__page_number__get",
        "parameters": [
          {
            "name": "page_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Page Number"
            }
          },
          {
            "name": "session_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Session Token"
            }
          },
          {
            "name": "project_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Project Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatHistoryResponsePage"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/InvalidOrExpiredSessionToken"
                    },
                    {
                      "$ref": "#/components/schemas/InvalidOrExpiredProjectToken"
                    }
                  ],
                  "title": "Response 401 Get Chat History Api Chat Get  Page Number  Get"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoAccessToGetQueryError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryNotFoundError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetrieveChatCompletionsError"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatCompletionErrorItem": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Chat completion for model failed"
          },
          "model_name": {
            "type": "string",
            "title": "Model Name",
            "description": "Model name that failed"
          }
        },
        "type": "object",
        "required": [
          "model_name"
        ],
        "title": "ChatCompletionErrorItem",
        "description": "Represents a failed attempt to generate a chat completion."
      },
...
      "ChatHistoryItem": {
        "properties": {
          "user_message": {
            "$ref": "#/components/schemas/ChatMessageItem",
            "description": "User's input message"
          },
          "assistant_responses": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ChatCompletionItem"
                },
                {
                  "$ref": "#/components/schemas/ChatCompletionErrorItem"
                }
              ]
            },
            "type": "array",
            "title": "Assistant Responses",
            "description": "Assistant's model responses"
          }
        },
        "type": "object",
        "required": [
          "user_message",
          "assistant_responses"
        ],
        "title": "ChatHistoryItem",
        "description": "Represents a single user-assistant interaction in a chat session."
      },
      "ChatHistoryResponse": {
        "properties": {
          "chat_id": {
            "type": "string",
            "format": "uuid",
            "title": "Chat Id",
            "description": "Unique chat session ID"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Title of the chat"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/ChatHistoryItem"
            },
            "type": "array",
            "title": "History",
            "description": "List of message exchanges"
          }
        },
        "type": "object",
        "required": [
          "chat_id",
          "title",
          "history"
        ],
        "title": "ChatHistoryResponse",
        "description": "Complete chat conversation history for a session."
      },
      "ChatHistoryResponsePage": {
        "properties": {
          "page": {
            "type": "integer",
            "title": "Page",
            "description": "Current page index"
          },
          "per_page": {
            "type": "integer",
            "title": "Per Page",
            "description": "Number of items per page"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ChatHistoryResponse"
            },
            "type": "array",
            "title": "Items",
            "description": "Paginated chat sessions"
          }
        },
        "type": "object",
        "required": [
          "page",
          "per_page",
          "items"
        ],
        "title": "ChatHistoryResponsePage",
        "description": "Paginated response containing multiple chat histories."
      },
...
      "CompletionErrorItem": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Completion for model failed"
          },
          "model_name": {
            "type": "string",
            "title": "Model Name",
            "description": "Model name that failed"
          }
        },
        "type": "object",
        "required": [
          "model_name"
        ],
        "title": "CompletionErrorItem",
        "description": "Represents an error during code completion."
      },
      "CompletionFeedbackPostResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Feedback recorded successfully."
          },
          "data": {
            "$ref": "#/components/schemas/ResponseFeedbackResponseData",
            "description": "Information about the recorded feedback"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "CompletionFeedbackPostResponse"
      },
      "CompletionPostResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Completions generated successfully. Ready for display in your IDE."
          },
          "data": {
            "$ref": "#/components/schemas/ResponseCompletionResponseData",
            "description": "Generated code completions including query ID and suggestions"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "CompletionPostResponse"
      },
...
      "ErrorResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Error message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ErrorResponse"
      },
      "FeedbackCompletion": {
        "properties": {
          "meta_query_id": {
            "type": "string",
            "format": "uuid",
            "title": "Meta Query Id",
            "description": "Meta Query ID"
          },
          "model_id": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Model Id",
            "description": "Model ID"
          },
          "was_accepted": {
            "type": "boolean",
            "title": "Was Accepted",
            "description": "Whether completion was accepted"
          },
          "ground_truth": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ground Truth",
            "description": "Actual code if available"
          }
        },
        "type": "object",
        "required": [
          "meta_query_id",
          "model_id",
          "was_accepted"
        ],
        "title": "FeedbackCompletion"
      },
      "FeedbackRecordingError": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Server failed to record feedback."
          }
        },
        "type": "object",
        "title": "FeedbackRecordingError"
      },
...
      "GenerationNotFoundError": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Generation record not found"
          }
        },
        "type": "object",
        "title": "GenerationNotFoundError"
      },
...
      "InvalidOrExpiredProjectToken": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Invalid or expired project token. Acquire a session token first."
          }
        },
        "type": "object",
        "title": "InvalidOrExpiredProjectToken"
      },
      "InvalidOrExpiredSessionToken": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Session not found! You are not authenticated or your session has expired. Login before you can perform this action."
          }
        },
        "type": "object",
        "title": "InvalidOrExpiredSessionToken"
      },
...
      "NoAccessToGetQueryError": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "default": "You do not have permission to access this query."
          }
        },
        "type": "object",
        "title": "NoAccessToGetQueryError"
      },
      "NoAccessToProvideFeedbackError": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "default": "You are not allowed to provide feedback for this query."
          }
        },
        "type": "object",
        "title": "NoAccessToProvideFeedbackError"
      },
...
      "QueryNotFoundError": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Query not found"
          }
        },
        "type": "object",
        "title": "QueryNotFoundError"
      },
...
      "ResponseCompletionItem": {
        "properties": {
          "model_id": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Model Id",
            "description": "Model ID"
          },
          "model_name": {
            "type": "string",
            "title": "Model Name",
            "description": "Model name"
          },
          "completion": {
            "type": "string",
            "title": "Completion",
            "description": "Generated code"
          },
          "generation_time": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Generation Time",
            "description": "Generation time in ms"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence",
            "description": "Confidence score of the result"
          }
        },
        "type": "object",
        "required": [
          "model_id",
          "model_name",
          "completion",
          "generation_time",
          "confidence"
        ],
        "title": "ResponseCompletionItem",
        "description": "Represents a successful model completion result."
      },
      "ResponseCompletionResponseData": {
        "properties": {
          "meta_query_id": {
            "type": "string",
            "format": "uuid",
            "title": "Meta Query Id",
            "description": "Meta Query ID"
          },
          "completions": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ResponseCompletionItem"
                },
                {
                  "$ref": "#/components/schemas/CompletionErrorItem"
                }
              ]
            },
            "type": "array",
            "title": "Completions",
            "description": "List of generated completions or errors"
          }
        },
        "type": "object",
        "required": [
          "meta_query_id",
          "completions"
        ],
        "title": "ResponseCompletionResponseData",
        "description": "Container for completion responses for a given meta-query."
      },
      "ResponseFeedbackResponseData": {
        "properties": {
          "meta_query_id": {
            "type": "string",
            "format": "uuid",
            "title": "Meta Query Id",
            "description": "Meta Query ID"
          },
          "model_id": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Model Id",
            "description": "Model ID"
          }
        },
        "type": "object",
        "required": [
          "meta_query_id",
          "model_id"
        ],
        "title": "ResponseFeedbackResponseData",
        "description": "Response model for storing feedback about completions."
      },
...
      "RetrieveChatCompletionsError": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Server failed to retrieve chat completions."
          }
        },
        "type": "object",
        "title": "RetrieveChatCompletionsError"
      },
      "RetrieveCompletionsError": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Server failed to retrieve completions."
          }
        },
        "type": "object",
        "title": "RetrieveCompletionsError"
      },
...
    }
  }
}